P0-b STAGE 2: JACKAL Hunter shadow_validate wiring
shared/contracts/risk.py:
- RiskDecision.final_score: int | None → float | None
- 운영 표현 정합 (probability + historical adjustment 후 소수점)
- analyst_score, devil_score는 int 유지 (Hunter int(...) 경로)

apps/jackal/risk_projection.py (신규):
- project_hunter_to_risk_decision()
- source_system="jackal_hunter", decision_stage="final"
- Devil/Final dict → RiskDecision payload
- safe access, bool coercion, None safe

apps/jackal/hunter.py:
- _shadow_validate_hunter_risk_decision() helper 추가
- _stage4_full_analysis() _apply_historical_context() 직후 호출
- shadow_validate(..., on_error="warn", context="jackal_hunter.stage4.risk_decision")
- 전체 try/except fail-open

tests:
- test_contracts_risk: float 검증 +4
- test_risk_projection (신규): 3 OK
- test_hunter_shadow_validate (신규): 3 OK

shadow mode only, enforced 전환 X
Hunter runtime behavior 무변경, hunt_log 무변경

contract_shadow_audit.log에 RiskDecision entries 누적 시작:
- 1-2주 audit 데이터 base STAGE 3 (Scanner) + enforced 결정

1133 tests OK, 회귀 0

+++ shared/contracts/risk.py
@@ -29,7 +29,7 @@ class RiskDecision(ContractModel):
     is_dead_cat: bool = False
     structural_decline: bool = False
 
-    final_score: int | None = Field(default=None, ge=0, le=100)
+    final_score: float | None = Field(default=None, ge=0, le=100)
     is_entry: bool = False
     block_reason: str | None = None
     decision_label: str | None = None
